aegix_project/requirements.txt
flask
python-dotenv
telethon
pyaes
rsa
pyasn1
requests
pytz
nltk
yfinance
quandl
pandas
numpy
matplotlib
seaborn
fpdf2
scikit-learn
geopy
beautifulsoup4
lxml
tqdm
plotly
dash
werkzeug
urllib3
scikit-learn